convert the modules to package
