+++ nyscforum/models.py
@@ -1,12 +1,6 @@
-from flask import Flask, render_template, url_for, flash, redirect
-from flask_sqlalchemy import SQLAlchemy
-from forms import RegistrationForm, LoginForm
 from datetime import datetime
+from nyscforum import db
 
-app = Flask(__name__)
-app.config['SECRET_KEY'] = '8639f43b71b224ad036d1047322a6b39'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///nysc.db'
-db = SQLAlchemy(app)
 
 class User(db.Model):
     
@@ -46,62 +40,7 @@ class Post(db.Model):
     title = db.Column(db.String(200), nullable=False)
     date_posted = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
     content = db.Column(db.Text, nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey(users.id), nullable=False)
+    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
 
     def __repr__(self):
         return "Post('{}', '{}')".format(self.title, self.date_posted)
-
-posts = [
-    {
-        'author': 'Victor Abayomi',
-        'title': 'NYSC Forum Post 1',
-        'date_posted': 'June 11, 2018',
-        'content': 'First NYSC Forum post content'
-    },
-    {
-        'author': 'Opeyemi Emmanuel',
-        'title': 'NYSC Forum Post 2',
-        'date_posted': 'June 30, 2018',
-        'content': 'Second NYSC Forum post content'
-    },
-    {
-        'author': 'Name Susan',
-        'title': 'NYSC Forum Post 3',
-        'date_posted': 'July 12, 2018',
-        'content': 'Third NYSC Forum post content'
-    }
-]
-
-@app.route("/")
-@app.route("/home")
-def home():
-    return render_template('home.html', posts=posts)
-
-
-@app.route("/about")
-def about():
-    return render_template('about.html', title='About')
-
-@app.route("/register", methods=['GET', 'POST'])
-def register():
-    form = RegistrationForm()
-
-    if form.validate_on_submit():
-        flash('Account created for {RegistrationForm.full_name.data} successfully!', 'success')
-        return redirect(url_for('home'))
-    return render_template('register.html', title='Register', form=form)
-
-@app.route("/login", methods=['GET', 'POST'])
-def login():
-    form = LoginForm()
-    
-    if form.validate_on_submit():
-        if form.email.data == 'ritymontero@gmail.com' and form.password.data == 'password':
-            flash('Login Successful!', 'success')
-            return redirect(url_for('home'))
-        else:
-            flash('Email or Password incorrect', 'danger')
-    return render_template('login.html', title='Login', form=form)
-
-if __name__ == '__main__':
-    app.run(debug=True)
