+++ nyscforum/routes.py
@@ -1,0 +1,56 @@
+from flask import render_template, url_for, flash, redirect
+from nyscforum import app
+from nyscforum.forms import RegistrationForm, LoginForm
+from nyscforum.models import User, Post
+
+posts = [
+    {
+        'author': 'Victor Abayomi',
+        'title': 'NYSC Forum Post 1',
+        'date_posted': 'June 11, 2018',
+        'content': 'First NYSC Forum post content'
+    },
+    {
+        'author': 'Opeyemi Emmanuel',
+        'title': 'NYSC Forum Post 2',
+        'date_posted': 'June 30, 2018',
+        'content': 'Second NYSC Forum post content'
+    },
+    {
+        'author': 'Name Susan',
+        'title': 'NYSC Forum Post 3',
+        'date_posted': 'July 12, 2018',
+        'content': 'Third NYSC Forum post content'
+    }
+]
+
+@app.route("/")
+@app.route("/home")
+def home():
+    return render_template('home.html', posts=posts)
+
+
+@app.route("/about")
+def about():
+    return render_template('about.html', title='About')
+
+@app.route("/register", methods=['GET', 'POST'])
+def register():
+    form = RegistrationForm()
+
+    if form.validate_on_submit():
+        flash('Account created for {RegistrationForm.full_name.data} successfully!', 'success')
+        return redirect(url_for('home'))
+    return render_template('register.html', title='Register', form=form)
+
+@app.route("/login", methods=['GET', 'POST'])
+def login():
+    form = LoginForm()
+    
+    if form.validate_on_submit():
+        if form.email.data == 'ritymontero@gmail.com' and form.password.data == 'password':
+            flash('Login Successful!', 'success')
+            return redirect(url_for('home'))
+        else:
+            flash('Email or Password incorrect', 'danger')
+    return render_template('login.html', title='Login', form=form)
